move refresh to completion contributor
